+++ CHANGELOG.md
@@ -56,6 +56,7 @@ replicas ([!166](https://storage.bsc.es/gitlab/hpc/gekkofs/-/merge_requests/141)
 - Modified write and reads to use a bitset instead of the traditional hash per chunk in the server.
 - Added reattemp support in get_fs_config to other servers, when the initial server fails.
 - Added support for Erasure codes ([!168]) using Jerasure lib and adding support for Read error injection.
+- Added support for on demand erasure codes calculation with `gkfs_ec_ondemand(fd)`
 ### New
 
 - Support for client-side per process logging, activated